Move DID to submodule
