+++ identity_core/src/did/mod.rs
@@ -1,0 +1,19 @@
+mod authentication;
+mod context;
+mod did;
+mod document;
+mod helpers;
+mod keys;
+mod parser;
+mod service;
+mod subject;
+mod utils;
+
+pub use authentication::Authentication;
+pub use context::Context;
+pub use did::{Param, DID};
+pub use document::DIDDocument;
+pub use keys::{KeyData, PublicKey, PublicKeyTypes};
+pub use service::{Service, ServiceEndpoint};
+pub use subject::Subject;
+pub use utils::{add_unique_to_vec, Dedup, HasId, IdCompare};
