+++ identity_core/src/did/subject.rs
@@ -1,11 +1,10 @@
-use crate::did::DID;
-
 use identity_diff::Diff;
 use serde::{Deserialize, Serialize};
-
 use std::{hash::Hash, str::FromStr};
 
-/// A wrapped `DID` type called a subject.  
+use crate::did::DID;
+
+/// A wrapped `DID` type called a subject.
 #[derive(Eq, PartialEq, Debug, Default, Clone, Serialize, Deserialize, Diff, Hash, PartialOrd, Ord)]
 #[serde(transparent)]
 #[diff(from_into)]
