+++ identity_core/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod common;
 pub mod did;
-pub mod did_parser;
-pub mod document;
-mod error;
-pub mod utils;
+pub mod error;
 
 pub use error::{Error, Result};
